renaming section locations component, adding categories
