+++ src/components/SectionCategories/SectionCategories.css
