+++ src/components/index.js
@@ -157,7 +157,7 @@ export { default as EditListingWizard } from './EditListingWizard/EditListingWiz
 export { default as Footer } from './Footer/Footer';
 export { default as SectionHero } from './SectionHero/SectionHero';
 export { default as SectionHowItWorks } from './SectionHowItWorks/SectionHowItWorks';
-export { default as SectionLocations } from './SectionLocations/SectionLocations';
+export { default as SectionCategories } from './SectionCategories/SectionCategories';
 export { default as SectionThumbnailLinks } from './SectionThumbnailLinks/SectionThumbnailLinks';
 export { default as Topbar } from './Topbar/Topbar';
 export { default as TopbarDesktop } from './TopbarDesktop/TopbarDesktop';
