+++ src/containers/LandingPage/LandingPage.js
@@ -10,7 +10,7 @@ import {
   Page,
   SectionHero,
   SectionHowItWorks,
-  SectionLocations,
+  SectionCategories,
   LayoutSingleColumn,
   LayoutWrapperTopbar,
   LayoutWrapperMain,
@@ -64,7 +64,7 @@ export const LandingPageComponent = props => {
           <ul className={css.sections}>
             <li className={css.section}>
               <div className={css.sectionContentFirstChild}>
-                <SectionLocations />
+                <SectionCategories />
               </div>
             </li>
             <li className={css.section}>
