+++ src/translations/de.json
@@ -707,8 +707,8 @@
   "SectionHowItWorks.part3Title": "3. Den/die Gastgeber(in) Bewerten",
   "SectionHowItWorks.titleLineOne": "Das Sauna-Ritual",
   "SectionHowItWorks.titleLineTwo": "(oder wie Saunatime funktioniert).",
-  "SectionLocations.listingsInLocation": "Saunas in {location}",
-  "SectionLocations.title": "Exotische Orte in Finnland erkunden",
+  "SectionCategories.listingsInLocation": "Saunas in {location}",
+  "SectionCategories.title": "Exotische Orte in Finnland erkunden",
   "SelectMultipleFilter.labelSelected": "{labelText} • {count}",
   "SelectMultipleFilterPlainForm.labelSelected": "{labelText} • {count}",
   "SelectSingleFilter.popupClear": "Löschen",
