+++ src/translations/en.json
@@ -707,8 +707,8 @@
   "SectionHowItWorks.part3Title": "3. Review your minder",
   "SectionHowItWorks.titleLineOne": "The öogo Way",
   "SectionHowItWorks.titleLineTwo": "(because the right-fit childcare is everything)",
-  "SectionLocations.listingsInLocation": "Minders in {location}",
-  "SectionLocations.title": "Find the Right-Fit Childcare",
+  "SectionCategories.listingsInCategory": "Find {category}",
+  "SectionCategories.title": "Find the Right-Fit Childcare",
   "SelectMultipleFilter.labelSelected": "{labelText} • {count}",
   "SelectMultipleFilterPlainForm.labelSelected": "{labelText} • {count}",
   "SelectSingleFilter.popupClear": "Clear",
