+++ src/translations/es.json
@@ -707,8 +707,8 @@
   "SectionHowItWorks.part3Title": "3. Deja una reseña",
   "SectionHowItWorks.titleLineOne": "El ritual del sauna",
   "SectionHowItWorks.titleLineTwo": "(o cómo funciona Saunatime)",
-  "SectionLocations.listingsInLocation": "Saunas en {location}",
-  "SectionLocations.title": "Explora lugares exóticos en Finlandia",
+  "SectionCategories.listingsInLocation": "Saunas en {location}",
+  "SectionCategories.title": "Explora lugares exóticos en Finlandia",
   "SelectMultipleFilter.labelSelected": "{labelText} • {count}",
   "SelectMultipleFilterPlainForm.labelSelected": "{labelText} • {count}",
   "SelectSingleFilter.popupClear": "Borrar",
