+++ src/translations/fr.json
@@ -707,8 +707,8 @@
   "SectionHowItWorks.part3Title": "3. Notez l'hôte",
   "SectionHowItWorks.titleLineOne": "Le rituel du sauna",
   "SectionHowItWorks.titleLineTwo": "(ou comment fonctionne Saunatime)",
-  "SectionLocations.listingsInLocation": "Saunas à {location}",
-  "SectionLocations.title": "Réchauffez-vous depuis une destination exotique en Finlande",
+  "SectionCategories.listingsInLocation": "Saunas à {location}",
+  "SectionCategories.title": "Réchauffez-vous depuis une destination exotique en Finlande",
   "SelectMultipleFilter.labelSelected": "{labelText} • {count}",
   "SelectMultipleFilterPlainForm.labelSelected": "{labelText} • {count}",
   "SelectSingleFilter.popupClear": "Effacer",
